chore: update sensor model and migration for decimal lat/lng

+++ migrations/20230406190735-sensor.js
@@ -24,11 +24,11 @@ module.exports = {
         },
       },
       latitude: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL(10, 8),
         allowNull: false
       },
       longitude: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL(11, 8),
         allowNull: false
       },
       isOn: {
